+++ src/body/params.ts
@@ -11,6 +11,17 @@ export class Params extends URLSearchParams {
     return [...this].length === 0;
   }
 
+  /*
+   * Create a new `Params` object with all empty values removed.
+   */
+  removeEmpty(): Params {
+    const params = new Params();
+    for (const [name, value] of this.entries()) {
+      if (value !== '') params.append(name, value);
+    }
+    return params;
+  }
+
   /**
    * Convert parameters into a plain object, useful for validation.
    */